Java class for evaluating fields length
This java class return the field length of a given set of CSV files.
Pass the folder that contains all the CSV files you want to analyze.
Syntax: java GetFieldsLenght [path to folder]

p.s. i know about the "Length" misspelling in the class name, but I'm lazy and I don't want to fix it, if you have more joy of life than me, do it.